requirements.txt
pyfiglet>=1.0.2
pillow>=10.3.0
colorama>=0.4.6
numpy>=2.0.0